Use torch.int instead of float.

+++ src/diffusers/models/unet_2d_condition.py
@@ -311,8 +311,8 @@ class UNet2DConditionModel(ModelMixin, ConfigMixin):
         if not torch.is_tensor(timesteps):
             # TODO: this requires sync between CPU and GPU. So try to pass timesteps as tensors if you can
             if sample.device.type == "mps":
-                # mps does not support int64 or float64
-                timesteps = torch.tensor([timesteps], dtype=torch.float32, device=sample.device)
+                # mps does not support 64-bit types
+                timesteps = torch.tensor([timesteps], dtype=torch.int32, device=sample.device)
             else:
                 timesteps = torch.tensor([timesteps], dtype=torch.long, device=sample.device)
         elif len(timesteps.shape) == 0:
